refactor(gateway): extract transport background tasks
